# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/attributes:
    get:
      tags:
        - Attributes
      summary: List Attributes
      description: Get a paginated list of Attributes with optional filtering and sorting
      operationId: listAttributes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Attributes
      summary: Create Attribute
      description: Create a new Attribute
      operationId: createAttribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttributeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/attribute_values:
    get:
      tags:
        - AttributeValues
      summary: List AttributeValues
      description: Get a paginated list of AttributeValues with optional filtering and sorting
      operationId: listAttributeValues
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeValueList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AttributeValues
      summary: Create AttributeValue
      description: Create a new AttributeValue
      operationId: createAttributeValue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttributeValueRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeValue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Attribute:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - attribute_type
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 100
        attribute_type:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateAttributeRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - attribute_type
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 100
        attribute_type:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    AttributeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AttributeValue:
      type: object
      required:
        - id
        - company_id
        - attribute_id
        - code
        - label
        - sort_order
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        attribute_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        label:
          type: string
          maxLength: 100
        label_en:
          nullable: true
          type: string
          maxLength: 100
        swatch_hex:
          nullable: true
          type: string
          maxLength: 10
        icon:
          nullable: true
          type: string
          maxLength: 255
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateAttributeValueRequest:
      type: object
      required:
        - company_id
        - attribute_id
        - code
        - label
        - sort_order
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        attribute_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        label:
          type: string
          maxLength: 100
        label_en:
          nullable: true
          type: string
          maxLength: 100
        swatch_hex:
          nullable: true
          type: string
          maxLength: 10
        icon:
          nullable: true
          type: string
          maxLength: 255
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    AttributeValueList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AttributeValue'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []